Refactor: Restructure project directories for better organization (#3)
* Initial project setup with FastAPI and common directory structure

This commit lays the foundational structure for our new Secure Storage Service POC.  The key components of this setup include:

1. **FastAPI Integration:**
   - FastAPI is chosen as the web framework due to its high performance, easy-to-use routing, and support for asynchronous programming, which is essential for scalability.

2. **Pipenv for Dependency Management:**
   - Pipenv is used to manage project dependencies, providing a clear separation of development and production packages, and ensuring consistent environments across different setups.

3. **Project Directory Structure:**
   - Established a standard directory structure to organize the codebase effectively. This includes:
     - `/models`: For database models.
     - `/routers`: For FastAPI routers to manage different endpoints.
     - `/services`: For business logic and service layer code.
     - `/config`: For app configuration
     - `/validation`: Validation logig for incoming requests
     - `/tests`: For unit and integration tests.

4. **Reasoning:**
   - This structure is set up in order to ease future development. It promotes clean separation of concerns, making the codebase easier to navigate, maintain, and scale. It also sets a clear path for future feature additions and enhancements.


I have updated the  README for detailed instructions on setting up and running the project.

Given this is a POC it will always remain on this feature branch, with that in mind we should branch off this in order to add features to this feature branch

* What:

Adding S3 to the project using localstack. Localstack is a emulation service used to emulate AWS services on our local environment.

Why:

As part of our service the ss poc will need a storage service to save files this part is simply setting up the Localstack.

* Updated README

* Updated README with docker instructions

* Updated README with docker instructions

* Updated README structure and added further instructions

* Updated README structure and added further instructions

* Updated README structure and added further instructions

* Updated README structure and added further instructions

* Refactor: Restructure project directories for better organization

Moved the 'config', 'models', 'validation', and 'services' folders under a new 'src' directory.
This change is aimed at improving the project structure by separating source code from test suites.
Such restructuring aligns with common Python project conventions, aiding in maintainability and clarity,
especially as the project scales. The 'test' directory remains at the root level to maintain a clear distinction
between source code and tests.

---------

Co-authored-by: MohammedHassanLatif93 <mohammed.latif@digital.justice.gov.uk>
Co-authored-by: Mohammed Hassan Latif <106679087+MohammedHassanLatif@users.noreply.github.com>