SDS-22 Ading dynamo db configuration
As part of the vconfiguration options we want to have a dynamc setup where we
pull the provider configuration from dynamo db.  This is only a spike of this and
does not involve a full fledged configuration.  The things that are not covered here is
how clietns will get this configurationito dynamo db.  Nor is this spike a solution to the configuration
problem, we may look at other options for storing configuration depending on what is convenient.
One of the alternatives is to use param store in AWS.